feat: modified naming,theming,icons structure
