+++ ReactDemo/my-react/src/Theme/index.tsx
@@ -1,0 +1,15 @@
+// import React from "react";
+import { createTheme } from "@mui/material";
+
+const theme = createTheme({
+    typography: {
+        subtitle1: {
+            color: '#373C38',
+            lineHeight: 1.5,
+        },
+        
+        
+    },
+});
+
+export default theme
