+++ ReactDemo/my-react/src/components/atoms/Image/index.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 interface ImageProp extends React.ImgHTMLAttributes<HTMLImageElement>{}
 
-const ImageComponent : React.FC<ImageProp> = ({...props}) => {
+const Image : React.FC<ImageProp> = ({...props}) => {
     return(
         <img {...props} />
     )
 }
 
-export default ImageComponent
+export default Image
