+++ ReactDemo/my-react/src/components/molecules/FooterImage/index.css
@@ -2,6 +2,7 @@
     text-align: start ;
     padding-left: 9px ;
     line-height: 0.5px;
+    color: #373C38;
     
 }
 
@@ -18,6 +19,7 @@
 .time{
     padding-top: 5px;
     padding-left: 50px;
+    color: #373C38;
     
     
 }
