+++ ReactDemo/my-react/src/components/molecules/FooterImage/index.tsx
@@ -1,42 +1,33 @@
 import React from "react";
-import bike from '../../../../public/assets/bike.svg';
-import bus from '../../../../public/assets/bus.svg';
-import car from '../../../../public/assets/car.svg';
-import train from '../../../../public/assets/train.svg';
-import { Stack, ThemeProvider, createTheme } from "@mui/material";
-import ImageComponent from "../../atoms/Image";
-import TextComponent from "../../atoms/Typography";
+import bike from '../../../../public/assets/icons/bike.svg';
+import bus from '../../../../public/assets/icons/bus.svg';
+import car from '../../../../public/assets/icons/car.svg';
+import train from '../../../../public/assets/icons/train.svg';
+import { Stack } from "@mui/material";
+import Image from "../../atoms/Image";
+import Typography from "../../atoms/Typography";
 import './index.css'; 
 
-const theme = createTheme({
-    typography: {
-        caption: {
-            color: '#373C38',
-        },
-    },
-});
 
 const ImageCard: React.FC = () => {
     return (
-        <ThemeProvider theme={theme}>
             <Stack className="container">
                 <Stack className="title">
-                    <TextComponent variant='caption'>Commute routes available :</TextComponent>
+                    <Typography variant='caption'>Commute routes available :</Typography>
                 </Stack>
                 <Stack className="image-row">
                     <Stack className="image-container">
-                        <ImageComponent src={bike} alt='bike' />
-                        <ImageComponent src={bus} alt='bus' />
-                        <ImageComponent src={car} alt='car' />
-                        <ImageComponent src={train} alt='train' />
+                        <Image src={bike} alt='bike' />
+                        <Image src={bus} alt='bus' />
+                        <Image src={car} alt='car' />
+                        <Image src={train} alt='train' />
                     </Stack>
                     <Stack className="time">
-                        <TextComponent variant='caption' >1hr 36 mins ago</TextComponent>   
+                        <Typography variant='caption' >1hr 36 mins ago</Typography>   
                     </Stack>
                     
                 </Stack>
             </Stack>
-        </ThemeProvider>
     );
 }
