+++ ReactDemo/my-react/src/components/template/TemplatePage/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Main from '../../organism/Main';
+import HpCard from '../../organism/HpCard';
 import { Stack } from '@mui/material';
 
 const Template: React.FC = () => {
     return (
         <Stack>
-            <Main />
+            <HpCard />
         </Stack>
     );
 };
