+++ ReactDemo/my-react/src/pages/Page/index.tsx
@@ -1,8 +1,14 @@
 import React from "react";
 import TemplatePage from "../../components/template/TemplatePage";
+import { ThemeProvider } from "@emotion/react";
+import theme from '../../../src/Theme'
+
 const FinalPage : React.FC = () => {
     return(
-        <TemplatePage />
+        <ThemeProvider theme={theme}>
+            <TemplatePage />
+        </ThemeProvider>
+        
     )
 }
 export default FinalPage
